move tensor types under `types/` submodule
