+++ vortex-tensor/src/lib.rs
@@ -11,19 +11,21 @@ use vortex_array::scalar_fn::session::ScalarFnSessionExt;
 use vortex_array::session::ArraySessionExt;
 use vortex_session::VortexSession;
 
-use crate::fixed_shape::FixedShapeTensor;
 use crate::scalar_fns::cosine_similarity::CosineSimilarity;
 use crate::scalar_fns::inner_product::InnerProduct;
 use crate::scalar_fns::l2_denorm::L2Denorm;
 use crate::scalar_fns::l2_norm::L2Norm;
 use crate::scalar_fns::sorf_transform::SorfTransform;
-use crate::vector::Vector;
+use crate::types::fixed_shape::FixedShapeTensor;
+use crate::types::vector::Vector;
 
 pub mod matcher;
 pub mod scalar_fns;
 
-pub mod fixed_shape;
-pub mod vector;
+mod types;
+
+pub use types::fixed_shape;
+pub use types::vector;
 
 pub mod encodings;
