+++ vortex-tensor/src/types/fixed_shape/metadata.rs
