+++ vortex-tensor/src/types/fixed_shape/proto.rs
@@ -8,7 +8,7 @@ use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_err;
 
-use crate::fixed_shape::FixedShapeTensorMetadata;
+use crate::types::fixed_shape::FixedShapeTensorMetadata;
 
 /// Protobuf representation of [`FixedShapeTensorMetadata`].
 ///
