+++ vortex-tensor/src/types/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+//! Internal homes for tensor extension types.
+
+pub mod fixed_shape;
+pub mod vector;
